requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==41.0.0
daphne==4.0.0
Django==4.2
django-filter==23.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
drf-spectacular==0.26.2
factory-boy==3.2.1
Faker==18.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
uritemplate==4.1.1
zope.interface==6.0
